Add "Clear Completed" button in footer
This patch adds the "Clear Completed" button in the footer, as specified the button
is only shown when at least one completed todo is present. Like the destroy todo
button, a`button_to` ActionView helper is used. A custom collection route on the
`todos` resource to handle the `clear_completed` action.

todoapp specification:
- https://github.com/tastejs/todomvc/blob/master/app-spec.md#clear-completed-button

Documentation:
- https://guides.rubyonrails.org/routing.html#adding-collection-routes
- https://api.rubyonrails.org/classes/ActiveRecord/Relation.html#method-i-delete_all
- https://api.rubyonrails.org/classes/ActiveRecord/Relation.html#method-i-destroy_all
- https://api.rubyonrails.org/classes/ActionView/Helpers/UrlHelper.html#method-i-button_to